requirements.txt
python-dotenv
google-api-python-client
google-auth
google-auth-oauthlib
openai
pydantic
pillow
